Fix a few mistakes in the setup instructions
The scope requested for Gitlab is not `api`, but rather should be `read_api` (as can be seen [here](https://github.com/getmeli/meli/blob/b045ce50d944320d515b3193340786f0fc8f129c/server/src/auth/passport/gitlab.ts#L32)).  Also, the value for `MELI_GITLAB_URL` should be the URL to the Gitlab instance to be authenticated against.  Unless you are running your own Gitlab instance, that would be `http://gitlab.com`.  The previous value in this document was probably a private Gitlab instance running on port 3003 used for testing or something.  But it wasn't a useful value for anybody reading the documentation.  This document assumes `gitlab.com` but points out how to handle the case of a private instance.